requirements.txt
streamlit>=1.32.0
pandas>=2.2.2
numpy>=2.0.0
matplotlib>=3.9.0
plotly>=5.20.0
requests>=2.31.0
python-dotenv>=1.0.1
scipy>=1.14.0
web3>=6.0.0
